backend/requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
sqlparse==0.5.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
requests>=2.31.0
Pillow>=10.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
